@@ -1,5 +1,7 @@
 module cosmossdk.io/indexer/postgres/testing
 
+go 1.23
+
 require (
 	cosmossdk.io/indexer/postgres v0.0.0-00010101000000-000000000000
 	cosmossdk.io/schema v0.1.1
@@ -31,5 +33,3 @@ require (
 replace cosmossdk.io/indexer/postgres => ../.
 
 replace cosmossdk.io/schema => ../../../schema
-
-go 1.22
