@@ -1,7 +1,7 @@
 module cosmossdk.io/server/v2/appmanager
 
-go 1.21
+go 1.23
 
 replace cosmossdk.io/core => ../../../core
 
-require cosmossdk.io/core v0.12.0
+require cosmossdk.io/core v1.0.0
