@@ -3,9 +3,9 @@ package runtime
 import (
 	"encoding/json"
 	"errors"
+	"slices"
 
 	gogoproto "github.com/cosmos/gogoproto/proto"
-	"golang.org/x/exp/slices"
 
 	runtimev2 "cosmossdk.io/api/cosmos/app/runtime/v2"
 	"cosmossdk.io/core/legacy"
@@ -77,6 +77,11 @@ func (a *App[T]) LoadLatest() error {
 // LoadHeight loads a particular height
 func (a *App[T]) LoadHeight(height uint64) error {
 	return a.db.LoadVersion(height)
+}
+
+// LoadLatestHeight loads the latest height.
+func (a *App[T]) LoadLatestHeight() (uint64, error) {
+	return a.db.GetLatestVersion()
 }
 
 // Close is called in start cmd to gracefully cleanup resources.
