@@ -5,11 +5,12 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"maps"
 	"reflect"
+	"slices"
 	"sort"
 
 	gogoproto "github.com/cosmos/gogoproto/proto"
-	"golang.org/x/exp/maps"
 	"google.golang.org/grpc"
 	proto "google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
@@ -41,7 +42,7 @@ func NewModuleManager[T transaction.Tx](
 	modules map[string]appmodulev2.AppModule,
 ) *MM[T] {
 	// good defaults for the module manager order
-	modulesName := maps.Keys(modules)
+	modulesName := slices.Sorted(maps.Keys(modules))
 	if len(config.PreBlockers) == 0 {
 		config.PreBlockers = modulesName
 	}
@@ -203,16 +204,13 @@ func (m *MM[T]) ExportGenesisForModules(
 		return nil, err
 	}
 
-	type genesisResult struct {
-		bz  json.RawMessage
-		err error
-	}
-
 	type ModuleI interface {
 		ExportGenesis(ctx context.Context) (json.RawMessage, error)
 	}
 
-	channels := make(map[string]chan genesisResult)
+	genesisData := make(map[string]json.RawMessage)
+
+	// TODO: make async export genesis https://github.com/cosmos/cosmos-sdk/issues/21303
 	for _, moduleName := range modulesToExport {
 		mod := m.modules[moduleName]
 		var moduleI ModuleI
@@ -221,27 +219,16 @@ func (m *MM[T]) ExportGenesisForModules(
 			moduleI = module.(ModuleI)
 		} else if module, hasABCIGenesis := mod.(appmodulev2.HasGenesis); hasABCIGenesis {
 			moduleI = module.(ModuleI)
+		} else {
+			continue
 		}
 
-		channels[moduleName] = make(chan genesisResult)
-		go func(moduleI ModuleI, ch chan genesisResult) {
-			jm, err := moduleI.ExportGenesis(ctx)
-			if err != nil {
-				ch <- genesisResult{nil, err}
-				return
-			}
-			ch <- genesisResult{jm, nil}
-		}(moduleI, channels[moduleName])
-	}
-
-	genesisData := make(map[string]json.RawMessage)
-	for moduleName := range channels {
-		res := <-channels[moduleName]
-		if res.err != nil {
-			return nil, fmt.Errorf("genesis export error in %s: %w", moduleName, res.err)
+		res, err := moduleI.ExportGenesis(ctx)
+		if err != nil {
+			return nil, err
 		}
 
-		genesisData[moduleName] = res.bz
+		genesisData[moduleName] = res
 	}
 
 	return genesisData, nil
@@ -438,7 +425,14 @@ func (m *MM[T]) RegisterServices(app *App[T]) error {
 			}
 		}
 
-		// TODO: register pre and post msg
+		// register pre and post msg
+		if module, ok := module.(appmodulev2.HasPreMsgHandlers); ok {
+			module.RegisterPreMsgHandlers(app.msgRouterBuilder)
+		}
+
+		if module, ok := module.(appmodulev2.HasPostMsgHandlers); ok {
+			module.RegisterPostMsgHandlers(app.msgRouterBuilder)
+		}
 	}
 
 	return nil
