@@ -89,7 +89,24 @@ func (a AppManager[T]) InitGenesis(
 
 // ExportGenesis exports the genesis state of the application.
 func (a AppManager[T]) ExportGenesis(ctx context.Context, version uint64) ([]byte, error) {
-	bz, err := a.exportGenesis(ctx, version)
+	zeroState, err := a.db.StateAt(version)
+	if err != nil {
+		return nil, fmt.Errorf("unable to get latest state: %w", err)
+	}
+
+	bz := make([]byte, 0)
+	_, err = a.stf.RunWithCtx(ctx, zeroState, func(ctx context.Context) error {
+		if a.exportGenesis == nil {
+			return errors.New("export genesis function not set")
+		}
+
+		bz, err = a.exportGenesis(ctx, version)
+		if err != nil {
+			return fmt.Errorf("failed to export genesis state: %w", err)
+		}
+
+		return nil
+	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to export genesis state: %w", err)
 	}
