@@ -7,6 +7,7 @@ import (
 	"time"
 
 	gogoproto "github.com/cosmos/gogoproto/proto"
+	gogoprotoany "github.com/cosmos/gogoproto/types/any"
 
 	"cosmossdk.io/core/appmodule"
 	corecontext "cosmossdk.io/core/context"
@@ -145,9 +146,19 @@ func (k Keeper) DispatchActions(ctx context.Context, grantee sdk.AccAddress, msg
 		}
 
 		// no need to use the branch service here, as if the transaction fails, the transaction will be reverted
-		_, err = k.MsgRouterService.InvokeUntyped(ctx, msg)
+		resp, err := k.MsgRouterService.InvokeUntyped(ctx, msg)
 		if err != nil {
 			return nil, fmt.Errorf("failed to execute message %d; message %v: %w", i, msg, err)
+		}
+
+		msgRespAny, err := gogoprotoany.NewAnyWithCacheWithValue(resp)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create any for response %d; message %s: %w", i, gogoproto.MessageName(msg), err)
+		}
+
+		results[i], err = gogoproto.Marshal(msgRespAny)
+		if err != nil {
+			return nil, fmt.Errorf("failed to marshal response %d; message %s: %w", i, gogoproto.MessageName(msg), err)
 		}
 	}
 
