@@ -498,7 +498,7 @@ metadata example:
 
 			// Since the --from flag is not required on this CLI command, we
 			// ignore it, and just use the 1st proposer in the JSON file.
-			if prop.Proposers == nil || len(prop.Proposers) == 0 {
+			if len(prop.Proposers) == 0 {
 				return errors.New("no proposers specified in proposal")
 			}
 			err = cmd.Flags().Set(flags.FlagFrom, prop.Proposers[0])
