@@ -111,7 +111,7 @@ func (s *KeeperTestSuite) TestDefaultMintFn() {
 	err = s.mintKeeper.DefaultMintFn(types.DefaultInflationCalculationFn)(s.ctx, s.mintKeeper.Environment, &minter, "block", 0)
 	s.NoError(err)
 
-	// set a maxsupply and call again
+	// set a maxSupply and call again. totalSupply will be bigger than maxSupply.
 	params, err := s.mintKeeper.Params.Get(s.ctx)
 	s.NoError(err)
 	params.MaxSupply = math.NewInt(10000000000)
@@ -122,14 +122,16 @@ func (s *KeeperTestSuite) TestDefaultMintFn() {
 	s.NoError(err)
 
 	// modify max supply to be almost reached
-	// we tried to mint 2059stake but we only get to mint 2000stake
+	// modify blocksPerYear to mint 2053 coins per block
+	// we tried to mint 2053stake, but we only get to mint 2000stake
 	params, err = s.mintKeeper.Params.Get(s.ctx)
 	s.NoError(err)
 	params.MaxSupply = math.NewInt(100000000000 + 2000)
+	params.BlocksPerYear = 2434275
 	err = s.mintKeeper.Params.Set(s.ctx, params)
 	s.NoError(err)
 
-	s.bankKeeper.EXPECT().MintCoins(s.ctx, types.ModuleName, sdk.NewCoins(sdk.NewCoin("stake", math.NewInt(792)))).Return(nil)
+	s.bankKeeper.EXPECT().MintCoins(s.ctx, types.ModuleName, sdk.NewCoins(sdk.NewCoin("stake", math.NewInt(2000)))).Return(nil)
 	s.bankKeeper.EXPECT().SendCoinsFromModuleToModule(s.ctx, types.ModuleName, authtypes.FeeCollectorName, gomock.Any()).Return(nil)
 
 	err = s.mintKeeper.DefaultMintFn(types.DefaultInflationCalculationFn)(s.ctx, s.mintKeeper.Environment, &minter, "block", 0)
@@ -143,7 +145,7 @@ func (s *KeeperTestSuite) TestBeginBlocker() {
 	s.bankKeeper.EXPECT().MintCoins(s.ctx, types.ModuleName, sdk.NewCoins(sdk.NewCoin("stake", math.NewInt(792)))).Return(nil)
 	s.bankKeeper.EXPECT().SendCoinsFromModuleToModule(s.ctx, types.ModuleName, authtypes.FeeCollectorName, gomock.Any()).Return(nil)
 
-	// get minter (it should get modified aftwerwards)
+	// get minter (it should get modified afterwards)
 	minter, err := s.mintKeeper.Minter.Get(s.ctx)
 	s.NoError(err)
 
