@@ -49,7 +49,7 @@ type AppModule struct {
 }
 
 // NewAppModule creates a new AppModule object.
-// If the mintFn argument is nil, then the SDK's default minting function will be used.
+// If the mintFn argument is nil, then the default minting function will be used.
 func NewAppModule(
 	cdc codec.Codec,
 	keeper keeper.Keeper,
