@@ -31,7 +31,7 @@ func DefaultParams() Params {
 		InflationMax:        math.LegacyNewDecWithPrec(5, 2),
 		InflationMin:        math.LegacyNewDecWithPrec(0, 2),
 		GoalBonded:          math.LegacyNewDecWithPrec(67, 2),
-		BlocksPerYear:       uint64(60 * 60 * 8766 / 5), // assuming 5 second block times
+		BlocksPerYear:       uint64(60 * 60 * 8766 / 5), // assuming 5-second block times
 		MaxSupply:           math.ZeroInt(),             // assuming zero is infinite
 	}
 }
