@@ -39,19 +39,23 @@ type Keeper struct {
 	LastBalance               collections.Item[math.Int]
 }
 
+const (
+	errModuleAccountNotSet = "%s module account has not been set"
+)
+
 func NewKeeper(cdc codec.BinaryCodec, env appmodule.Environment, ak types.AccountKeeper, bk types.BankKeeper, sk types.StakingKeeper, authority string,
 ) Keeper {
 	// ensure pool module account is set
 	if addr := ak.GetModuleAddress(types.ModuleName); addr == nil {
-		panic(fmt.Sprintf("%s module account has not been set", types.ModuleName))
+		panic(fmt.Sprintf(errModuleAccountNotSet, types.ModuleName))
 	}
 	// ensure stream account is set
 	if addr := ak.GetModuleAddress(types.StreamAccount); addr == nil {
-		panic(fmt.Sprintf("%s module account has not been set", types.StreamAccount))
+		panic(fmt.Sprintf(errModuleAccountNotSet, types.StreamAccount))
 	}
 	// ensure protocol pool distribution account is set
 	if addr := ak.GetModuleAddress(types.ProtocolPoolDistrAccount); addr == nil {
-		panic(fmt.Sprintf("%s module account has not been set", types.ProtocolPoolDistrAccount))
+		panic(fmt.Sprintf(errModuleAccountNotSet, types.ProtocolPoolDistrAccount))
 	}
 
 	sb := collections.NewSchemaBuilder(env.KVStoreService)
@@ -172,6 +176,31 @@ func (k Keeper) SetToDistribute(ctx context.Context) error {
 
 	// Calculate the amount to be distributed
 	amountToDistribute := distributionBalance.Sub(lastBalance)
+
+	// Check if there are any recipients to distribute to, if not, send straight to the community pool and avoid
+	// setting the distributions
+	hasContinuousFunds := false
+	err = k.ContinuousFund.Walk(ctx, nil, func(_ sdk.AccAddress, _ types.ContinuousFund) (bool, error) {
+		hasContinuousFunds = true
+		return true, nil
+	})
+	if err != nil {
+		return err
+	}
+
+	// if there are no continuous funds, send all the funds to the community pool and reset the last balance
+	if !hasContinuousFunds {
+		poolCoins := sdk.NewCoins(sdk.NewCoin(denom, amountToDistribute))
+		if err := k.bankKeeper.SendCoinsFromModuleToModule(ctx, types.ProtocolPoolDistrAccount, types.ModuleName, poolCoins); err != nil {
+			return err
+		}
+
+		if !lastBalance.IsZero() { // only reset if the last balance is not zero (so we leave it at zero/nil)
+			return k.LastBalance.Set(ctx, math.ZeroInt())
+		}
+
+		return nil
+	}
 
 	if err = k.Distributions.Set(ctx, k.HeaderService.HeaderInfo(ctx).Time, amountToDistribute); err != nil {
 		return fmt.Errorf("error while setting Distributions: %w", err)
